third_party/blink/web_tests/virtual/mutation-events-enabled/dom/mutation-event-tests/fast/dom/shadow/suppress-mutation-events-in-shadow-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeInserted' mutation event. This event type is no longer supported, and will be removed from this browser VERY soon. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Test that mutation events are not triggered in shadow DOM

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutationEventTriggered is false
PASS successfullyParsed is true

TEST COMPLETE
